feat: 댓글 패키지 구축
- 기본적인 댓글 패키지 구축 완료

+++ src/main/java/com/westudy/post/mapper/PostDetailMapper.java
@@ -11,5 +11,5 @@ import java.util.List;
 public interface PostDetailMapper {
     PostDetailDBDTO findPostDetailById(@Param("postId") long postId);
     List<PostResponseDTO> findSearchPosts(String keyword, int size, int offset);
-    int countSearchPosts(String keyword);
+    long countSearchPosts(String keyword);
 }
